pyproject.toml
[project]
name = "poster-to-markdown"
version = "0.1.0"
description = "Convert research posters to markdown summaries using OpenAI"
authors = [
    { name = "Seung Hyun Kim", email = "skim0119@gmail.com" },
]
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "asyncio",
    "click",
    "mcp[cli]>=1.9.1",
    "openai",
    "Pillow",
    "pillow-heif",
    "python-dateutil",
]

[project.scripts]
poster-to-markdown = "poster_to_markdown.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 95
target-version = "py312"
lint.select = ["E", "F", "I", "N", "W"]
lint.ignore = []

[dependency-groups]
dev = [
    "ruff",
]
